Move Thrift methods out of MetastoreUtil
